test(e2e): add photo scenarios and metrics check
- 新增 photo 模块 30 个 e2e 场景(上传/列表/详情/下载/删除/相册/评论/点赞)
  与 /metrics 指标存在性断言; preprea 幂等清理 e2e 照片数据与 S3 孤儿对象
- 修复 server.build_info 的 commit 标签: build.rs 改用 GITHUB_SHA 与随提交
  变化的 ref 路径作为重建触发, 避免提交后仍报旧 commit
- 修掉 metrics feature 组合下的 useless_conversion / unnecessary_cast
- 新增手动触发的 e2e CI workflow

+++ server/src/setup/bases/metrics/redis.rs
@@ -7,7 +7,7 @@ use metrics::gauge;
 pub fn collect_redis_metrics(pool: &Pool) {
     let status = pool.status();
     // deadpool 的 `size` 为连接总数（含空闲），active 需扣除空闲，与 database 连接池语义对齐。
-    let idle = status.available as usize;
+    let idle = status.available;
     gauge!("redis.connections.active").set(status.size.saturating_sub(idle) as f64);
     gauge!("redis.connections.idle").set(idle as f64);
     gauge!("redis.connections.waiting").set(status.waiting as f64);
